proyect/requirements.txt
django
djangorestframework
django-cors-headers
pandas
scikit-learn
gunicorn